Refactor firebase folder structure
