+++ src/integrations/index.ts
@@ -1,4 +1,4 @@
-export * from './firebase/collectionHandlers';
-export * from './firebase/authHandlers';
-export * from './firebase/useAuthorizedUser';
-export * from './firebase/firebaseContext';
+export * from './firebase/handlers/collectionHandlers';
+export * from './firebase/handlers/authHandlers';
+export * from './firebase/hooks/useAuthorizedUser';
+export * from './firebase/contexts/firebaseContext';
